backend/requirements.txt
# Core Framework - Updated to latest stable versions (October 2025)
fastapi==0.119.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
pydantic-settings==2.11.0

# Database - Enhanced for multi-tenancy and performance
sqlalchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.0
psycopg2-binary==2.9.11
pgvector==0.4.1

# AI/ML - Updated to stable LangChain 1.x and latest providers
langchain==1.0.2
langchain-openai==1.0.1
langchain-anthropic==1.0.0
langchain-huggingface==0.1.0
openai==2.6.0
anthropic==0.71.0
google-generative-ai-python==1.0.0
huggingface-hub==0.30.0
sympy==1.14.0

# WebSocket - Enhanced for scalability
python-socketio==5.14.2
python-socketio[asyncio_client]==5.14.2

# Task Queue - Enhanced reliability
celery==5.5.3
redis==6.4.0

# HTTP Client - Performance improvements
httpx==0.28.1
aiohttp==3.13.1

# Authentication - Security hardened
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Monitoring - Enhanced observability
prometheus-client==0.23.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0

# Utilities - Enhanced for ELCA compliance
python-dotenv==1.1.1
structlog==25.4.0
tenacity==9.1.0

# ELCA-Specific Enhancements
python-i18n==0.1.0
cryptography==44.0.0
pydantic-extra-types==2.6.0

# Multi-tenancy support
sqlalchemy-utils==0.41.1
tenant-schemas==1.10.0

# Security and compliance
bandit==1.7.10
safety==3.2.0
# Production server
gunicorn==21.2.0
